fix(business) 权限校验（不同用户）+ 统一id为Long

+++ elm_bk/src/main/java/com/tju/elm_bk/controller/BusinessController.java
@@ -43,7 +43,7 @@ public class BusinessController {
     @GetMapping("/{id}")
     @Operation(summary = "根据id获取某店铺详情", description = "根据id获取某店铺详情")
 
-    public HttpResult<BusinessVO> getBusiness(@PathVariable("id") Integer id) {
+    public HttpResult<BusinessVO> getBusiness(@PathVariable("id") Long id) {
         if (id == null || id <= 0) {
             log.warn("获取店铺详情请求参数错误: id={}", id);
             throw new APIException(ResultCodeEnum.PARAM_NOT_MATCHED);
@@ -66,7 +66,7 @@ public class BusinessController {
      */
     @PutMapping("/{id}")
     @Operation(summary = "更新某店铺信息", description = "更新某店铺信息")
-    public HttpResult<BusinessVO> updateBusiness(@PathVariable("id") Integer id,@RequestBody BusinessUpdateDTO updateDto){
+    public HttpResult<BusinessVO> updateBusiness(@PathVariable("id") Long id,@RequestBody BusinessUpdateDTO updateDto){
         if (id == null || id <= 0) {
             log.warn("更新店铺信息请求参数错误: id={}", id);
             throw new APIException(ResultCodeEnum.PARAM_NOT_MATCHED);
@@ -78,7 +78,7 @@ public class BusinessController {
 
      @DeleteMapping("/{id}")
      @Operation(summary = "删除某店铺")
-    public HttpResult<BusinessVO> deleteBusiness(@PathVariable("id") Integer id) {
+    public HttpResult<BusinessVO> deleteBusiness(@PathVariable("id") Long id) {
         if (id == null || id <= 0) {
             log.warn("删除店铺信息请求参数错误: id={}", id);
             throw new APIException(ResultCodeEnum.PARAM_NOT_MATCHED);
@@ -88,7 +88,7 @@ public class BusinessController {
     }
      @PatchMapping("/{id}")
      @Operation(summary = "部分更新某店铺信息")
-    public HttpResult<BusinessVO> patchBusiness(@PathVariable("id") Integer id,@RequestBody BusinessUpdateDTO updateDto) {
+    public HttpResult<BusinessVO> patchBusiness(@PathVariable("id") Long id,@RequestBody BusinessUpdateDTO updateDto) {
         if (id == null || id <= 0) {
             log.warn("更新店铺信息请求参数错误: id={}", id);
             throw new APIException(ResultCodeEnum.PARAM_NOT_MATCHED);
